# Cache table refs