Stacked on #2439 

Closes #2080 

Rather than creating new references each time we access ``tables.edges``, this caches the objects in TableCollection. There were a few things that tripped me up along the way, but I think it does lead to simpler code and makes it fairly clear  now how we addess (part of, at least) making ``ts.tables`` return a view of the TableCollection data.

Benchmarks:
```python
import msprime
import sys
import time

before = time.perf_counter()
ts = msprime.sim_ancestry(
    100000,
    sequence_length=1000000,
    population_size=10_000,
    recombination_rate=1e-8,
    random_seed=1234,
)
ts = msprime.sim_mutations(ts, rate=1e-7, random_seed=1)
print(ts)
duration = time.perf_counter() - before
print(f"Simulation of {ts.num_trees} trees done after {duration:.2f} seconds")

tables = ts.tables
before = time.perf_counter()
for _ in range(10000):
    tables.nodes
    tables.edges
    tables.individuals
    tables.populations
    tables.mutations
    tables.sites
    tables.provenances
    tables.migrations
duration = time.perf_counter() - before
print(f"Pulled out tables in {duration:.2g} seconds")

before = time.perf_counter()
populations = tables.populations
for _ in range(10000):
    populations.metadata_schema
duration = time.perf_counter() - before
print(f"Pulled out metadata_schema in {duration:.2g} seconds")

before = time.perf_counter()
edges = tables.edges
for _ in range(10000):
    edges.left
duration = time.perf_counter() - before
print(f"Pulled out edges.left in {duration:.2g} seconds")

before = time.perf_counter()
for _ in range(1000):
    tables = ts.tables
    # tables.nodes
duration = time.perf_counter() - before
print(f"Pulled out ts.tables in {duration:.2g} seconds")

before = time.perf_counter()
for _ in range(1000):
    nodes = ts.tables.nodes
duration = time.perf_counter() - before
print(f"Pulled out ts.tables.nodes in {duration:.2g} seconds")
```
**NOTE: the last two run 10x fewer iterations as ``ts.tables`` is still slowwww.**

# Results

# Current upstream main
```
╔═══════════════════════════╗
║TreeSequence               ║
╠═══════════════╤═══════════╣
║Trees          │       4885║
╟───────────────┼───────────╢
║Sequence Length│    1000000║
╟───────────────┼───────────╢
║Time Units     │generations║
╟───────────────┼───────────╢
║Sample Nodes   │     200000║
╟───────────────┼───────────╢
║Total Size     │   32.5 MiB║
╚═══════════════╧═══════════╝
╔═══════════╤══════╤═════════╤════════════╗
║Table      │Rows  │Size     │Has Metadata║
╠═══════════╪══════╪═════════╪════════════╣
║Edges      │418790│ 12.8 MiB│          No║
╟───────────┼──────┼─────────┼────────────╢
║Individuals│100000│  2.7 MiB│          No║
╟───────────┼──────┼─────────┼────────────╢
║Migrations │     0│  8 Bytes│          No║
╟───────────┼──────┼─────────┼────────────╢
║Mutations  │ 51949│  1.8 MiB│          No║
╟───────────┼──────┼─────────┼────────────╢
║Nodes      │404076│ 10.8 MiB│          No║
╟───────────┼──────┼─────────┼────────────╢
║Populations│     1│224 Bytes│         Yes║
╟───────────┼──────┼─────────┼────────────╢
║Provenances│     2│  1.6 KiB│          No║
╟───────────┼──────┼─────────┼────────────╢
║Sites      │ 50650│  1.2 MiB│          No║
╚═══════════╧══════╧═════════╧════════════╝

Simulation of 4885 trees done after 1.26 seconds
Pulled out tables in 0.19 seconds
Pulled out metadata_schema in 0.0011 seconds
Pulled out edges.left in 0.99 seconds
Pulled out ts.tables in 3.4 seconds
Pulled out ts.tables.nodes in 4.5 seconds
```
![upsteam-main](https://user-images.githubusercontent.com/2664569/180824564-f018a05e-b135-44b9-92fb-3b92b0db65f5.png)

# This branch
```
Simulation of 4885 trees done after 1.30 seconds
Pulled out tables in 0.005 seconds
Pulled out metadata_schema in 0.0029 seconds
Pulled out edges.left in 1.1 seconds
Pulled out ts.tables in 3.5 seconds
Pulled out ts.tables.nodes in 7.4 seconds
```
![this-branch](https://user-images.githubusercontent.com/2664569/180825022-f188f1ef-18b6-44e3-a7b1-8ca19be6e41f.png)