@jeromekelleher 2022-07-25
Note: it's not clear why ``ts.tables.edges`` is 2X slower here than in main. Something to look into...