@jeromekelleher 2022-07-27
Current version means we parse the metadata schema every time we decode a row, so this will be very bad for row-access performance.