@jeromekelleher 2022-07-27
Actually, I think the performance of this is fine. Here's my benchmark:
```python
import time
import tskit
import msprime

demography = msprime.Demography.stepping_stone_model([100] * 100, migration_rate=0.0)
for k in range(100):
    demography[k].name = f"awesome_pop_{k}"
    demography[k].description = "This population is totally awesome"
    demography[k].extra_metadata = {f"x_{j}": list(range(j)) for j in range(100)}
ts = msprime.sim_ancestry({0: 2}, demography=demography, random_seed=1)

# print(ts.tables.populations)
before = time.perf_counter()
for _ in range(100):
    for pop in ts.populations():
        # print(pop.metadata)
        # x = pop.metadata
        pass
duration = time.perf_counter() - before

print(f"ts.populations in {duration:.2g} seconds")

# print(ts.tables.populations)
before = time.perf_counter()
for _ in range(100):
    for pop in ts.populations():
        # print(pop.metadata)
        x = pop.metadata
        pass
duration = time.perf_counter() - before
print(f"ts.populations decode metadata in {duration:.2g} seconds")

tables = ts.tables
# print(ts.tables.populations)
before = time.perf_counter()
for _ in range(100):
    for pop in tables.populations:
        # print(pop.metadata)
        # x = pop.metadata
        pass
duration = time.perf_counter() - before
print(f"tables.populations in {duration:.2g} seconds")

tables = ts.tables
# print(ts.tables.populations)
before = time.perf_counter()
for _ in range(100):
    for pop in tables.populations:
        # print(pop.metadata)
        x = pop.metadata
        pass
duration = time.perf_counter() - before
print(f"tables.populations decode metadata in {duration:.2g} seconds")
```
For upstream/main we get:
```
ts.populations in 0.02 seconds
ts.populations decode metadata in 2.7 seconds
tables.populations in 0.026 seconds
tables.populations decode metadata in 2.7 seconds
```
For this branch we get:
```
ts.populations in 0.02 seconds                                                                                                                                                                                                                
ts.populations decode metadata in 2.7 seconds                                                                                                                                                                                                 
tables.populations in 0.028 seconds                                                                                                                                                                                                           
tables.populations decode metadata in 2.7 seconds      
```
So, there's no difference really when we're decoding significant amounts of metadata. The reason it doesn't matter that we're not caching the parsed metadata schema is because we have a functools.lru_cache on the ``parse_metadata_schema`` function. So, we incur the cost of creating the unicode string and looking up that cache each time, but I bet that's not much different to any other mechanism we'd come up with, and it's a lot simpler.

What do you think @benjeffery - should we try other benchmarks here?