@benjeffery 2022-07-28
Ahhh, interesting. The LRU cache was added to improve replicate performance after the initial TableCollection metadata implementation that has the instance-level cache. I think we can completely remove the instance level caching with little penalty then. I'll review this now.